+++ src/lib/components/config/ColorConfig.svelte
@@ -11,18 +11,18 @@
 
 	let { entryId, locked }: Props = $props();
 
-	// we don't use hashtags in cfg files, since they are turned into comments
-	// but the color input works with the # prefix, so we have to separate them
 	let content = entryId.entry.value.content as string;
-	let hexCode = $state(`#${content}`);
+	let hasHashtag = content.startsWith('#');
+
+	let hexCode = $state(hasHashtag ? content : `#${content}`);
 
 	async function onReset(value: ConfigValue) {
-		hexCode = `#${value.content}`;
+		hexCode = hasHashtag ? (value.content as string) : `#${value.content}`;
 		await submit();
 	}
 
 	async function submit() {
-		content = hexCode.slice(1);
+		content = hasHashtag ? hexCode : hexCode.slice(1);
 
 		await setConfigEntry(entryId, {
 			type: 'string',