+++ src/lib/util.ts
@@ -128,7 +128,7 @@ export function fileToBase64(file: File): Promise<string> {
 }
 
 export function isValidHex(str: string) {
-	return /^([0-9A-Fa-f]{6})$/.test(str);
+	return /^#?([0-9A-Fa-f]{6})$/.test(str);
 }
 
 export function discordAvatarUrl(discordId: string, avatar: string) {